Separate simulation controls to standalone component
